+++ lib/src/common/typedefs.dart
@@ -23,10 +23,10 @@ typedef DataMapEntry = MapEntry<String, Object?>;
 /// Type definition for [DescriptionBuilder] simplifying descriptive landing page related widgets.
 ///
 ///
-/// [T] type of the [ThemeI] implementation that defines application base theme contract.
+/// [T] type of the [ThemeDataI] implementation that defines application base theme contract.
 ///
 ///
 /// [theme] the current [T] instance result.
 ///
 /// [foreColor] the preferred foreground color for the description.
-typedef DescriptionBuilder<T extends ThemeI> = InlineSpan Function(T theme, Color foreColor);
+typedef DescriptionBuilder<T extends ThemeDataI> = InlineSpan Function(T theme, Color foreColor);