+++ lib/src/utils/injector.dart
@@ -23,11 +23,4 @@ final class Injector {
   static T get<T extends Object>() {
     return _injector.get<T>();
   }
-
-  /// Gets the [ThemeManagerI] instance from the dependency injector.
-  static ThemeManagerI<TThemeB> getThemeManager<TThemeB extends ThemeI>() {
-    final ThemeManagerI<TThemeB> themeManager = _injector.get<ThemeManagerI<TThemeB>>();
-
-    return themeManager;
-  }
 }