+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # CHANGELOG
 
+## 7.0.0 - [23.06-2025]
+
+- Notes:
+
+  - Changed the theming concept from a static ThemeManagerI dependency injected to an InheritedWidget implementation calling ThemeManager.of(context).
+  
+  - Refactorized PackageLanding concept to match this new Theming management.
+
+- Dependencies:
+
+    | Package                                 | Previous Version | Current Version |
+    |:----------------------------------------|:----------------:|:---------------:|
+    | get_it                                  | 8.0.3            | 8.0.3           |
+    | go_router                               | 15.2.3           | 15.1.2          |
+    | device_info_plus                        | 11.5.0           | 11.4.0          |
+
+- Dev Dependencies:
+
+    | Package                                 | Previous Version | Current Version |
+    |:----------------------------------------|:----------------:|:---------------:|
+    | lints                                   | 6.0.0            | 6.0.0           |
+
 ## 6.1.1 - [25.05-2025]
 
 - Dependencies: